+++ host-configs/matrix-toss_4_x86_64_ib-gcc@13.3.1_cuda.cmake
@@ -1,16 +1,16 @@
 #------------------------------------------------------------------------------
 # !!!! This is a generated file, edit at own risk !!!!
 #------------------------------------------------------------------------------
-# CMake executable path: /usr/tce/bin/cmake
+# CMake executable path: /usr/tce/packages/cmake/cmake-3.26.3/bin/cmake
 #------------------------------------------------------------------------------
 
-set(CMAKE_PREFIX_PATH "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/blt-develop-vwop4eonp3ifr22723obz3qzpkhp7frz;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/c2c-1.8.0-xcseimmxzzygacwefcb4orsdjkexiave;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/caliper-git.527e1ae3a16487e425707e928e05c43d372e144f_master-yaa4lvtogjizlgwmk52mcunsucvykvkg;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/conduit-0.9.5-eegw7x7gkohlehrruiegywco4xiamktk;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/mfem-4.9.0-oeogieerz3o2odisl7obn44yzugviidm;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/raja-git.e827035c630e71a9358e2f21c2f3cf6fd5fb6605_develop-4shuz3tf6ej2nusu2p7gy3tub4fl7cit;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/umpire-git.0372fbd6e1f17d7e6dd72693f8b857f3ec7559e9_develop-mekv5ckdc5utbqgxefdhsaeuftidlnrz;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/adiak-0.4.0-527775ljxopt67ho3pvm5tezd37rmjc2;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/libunwind-1.8.3-74hjhhtejjsdgxjpp262xlm5ow6psq4r;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/hdf5-1.8.23-dznlaqnsupnxvdgp4gvkneu32tnd2ebk;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/parmetis-4.0.3-izp4ercwbua4ufxipqapaajpjwe22bdt;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/hypre-2.27.0-ceinfx2z45d7ap3i2lif6j2qw6jw7vhf;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/camp-git.a8caefa9f4c811b1a114b4ed2c9b681d40f12325_main-idirlfyyjti4dfn4lbutnn6tvabsyyzm;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/fmt-11.0.2-vn5l5jfzk37utfwbibtqwpkc2pyrvg3v;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/metis-5.1.0-hfskjufounxgma6wny7v2p3b2z7jxvsm;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/none-none/gcc-runtime-13.3.1-k43mnfzd4ehqvovn4d5qq7b5f6sfetsx;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/none-none/compiler-wrapper-1.0-32dgmtd4fhqezj4ib4pbwjnwrxcuzxs7;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/view/python-3.13.5;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/view/python-3.13.5;/collab/usr/gapps/shroud/public/toss_4_x86_64_ib/shroud-0.14.0;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/view/python-3.13.5;/usr/tce;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/gcc-13.3.1/cppcheck-2.18.0-d2tmb7ls2d5vgh6vhcyohecwcu6msopb;/usr/tce/packages/cuda/cuda-12.9.1;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/view/doxygen-1.13.2;/usr/tce/packages/gcc/gcc-13.3.1;/usr/tce/packages/clang/clang-19.1.3;/usr/tce/packages/mvapich2/mvapich2-2.3.7-gcc-13.3.1;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/view/python-3.13.5" CACHE STRING "")
+set(CMAKE_PREFIX_PATH "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/blt-develop-vkyjvzyg7n4ig7x4llo53lndvnwsrbjq;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/c2c-1.8.0-3bch3wbxcl3eji2wsbbbzmagowxe4lpc;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/caliper-git.9ab0e64d2e328d1fdff2ffebb53f30a899d6a598_master-ss6he6vnxctdifczfduoqdpfyjltt2r3;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/conduit-0.9.5-icskqkcmikkfw2itxiq6xdt6nilzlsvr;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/mfem-4.9.0-lkbiphsoccgpbaqo2cujuxcf3dqb4gga;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/raja-git.3b8b59a1e9be2e1066c0d77372b3bf5956e6d6e2_develop-2u2olu3wtrgskohwge6z2r5p4pbn24xi;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/umpire-git.0372fbd6e1f17d7e6dd72693f8b857f3ec7559e9_develop-qqnegxhkanugllghjqvoqjilj2h4ouuf;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/adiak-0.4.0-45h4d32dvvatilujstudj3ucvmdeyi6a;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/libunwind-1.8.3-74hjhhtejjsdgxjpp262xlm5ow6psq4r;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/hdf5-1.8.23-gokrordtweyfprtzgrzsycyilhy32uc7;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/parmetis-4.0.3-euggsyckxbtpzv6kzq74k3ndngpw2yqh;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/hypre-2.27.0-ceinfx2z45d7ap3i2lif6j2qw6jw7vhf;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/camp-git.a8caefa9f4c811b1a114b4ed2c9b681d40f12325_main-idk67733aarmoow76xcm5idyfijqjntn;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/fmt-11.0.2-faqgdwrlrl4nwv35msienhldhqrq2vce;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/metis-5.1.0-ovo3o7do7zr5fxw6mi2flzy6ftfsavk2;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/none-none/gcc-runtime-13.3.1-k43mnfzd4ehqvovn4d5qq7b5f6sfetsx;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/none-none/compiler-wrapper-1.0-32dgmtd4fhqezj4ib4pbwjnwrxcuzxs7;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/view/python-3.13.5;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/view/python-3.13.5;/collab/usr/gapps/shroud/public/toss_4_x86_64_ib/shroud-0.14.0;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/view/python-3.13.5;/usr/tce/packages/cmake/cmake-3.26.3;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/gcc-13.3.1/cppcheck-2.18.0-d2tmb7ls2d5vgh6vhcyohecwcu6msopb;/usr/tce/packages/cuda/cuda-12.9.1;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/view/doxygen-1.13.2;/usr/tce/packages/gcc/gcc-13.3.1;/usr/tce/packages/clang/clang-19.1.3;/usr/tce/packages/mvapich2/mvapich2-2.3.7-gcc-13.3.1;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib/2025_09_25_11_33_37/view/python-3.13.5" CACHE STRING "")
 
 set(CMAKE_INSTALL_RPATH_USE_LINK_PATH "ON" CACHE STRING "")
 
-set(CMAKE_BUILD_RPATH "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/axom-develop-u2lyxopctg45ibm3jwo5ej3fogit3ud2/lib;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/axom-develop-u2lyxopctg45ibm3jwo5ej3fogit3ud2/lib64;;/collab/usr/global/tools/tce4/packages/gcc/gcc-13.3.1/lib/gcc/x86_64-redhat-linux/13" CACHE STRING "")
+set(CMAKE_BUILD_RPATH "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/axom-develop-cgeagxjp7rbuciiqqnlu5ajev42355xw/lib;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/axom-develop-cgeagxjp7rbuciiqqnlu5ajev42355xw/lib64;;/collab/usr/global/tools/tce4/packages/gcc/gcc-13.3.1/lib/gcc/x86_64-redhat-linux/13" CACHE STRING "")
 
-set(CMAKE_INSTALL_RPATH "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/axom-develop-u2lyxopctg45ibm3jwo5ej3fogit3ud2/lib;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1/axom-develop-u2lyxopctg45ibm3jwo5ej3fogit3ud2/lib64;;/collab/usr/global/tools/tce4/packages/gcc/gcc-13.3.1/lib/gcc/x86_64-redhat-linux/13" CACHE STRING "")
+set(CMAKE_INSTALL_RPATH "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/axom-develop-cgeagxjp7rbuciiqqnlu5ajev42355xw/lib;/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1/axom-develop-cgeagxjp7rbuciiqqnlu5ajev42355xw/lib64;;/collab/usr/global/tools/tce4/packages/gcc/gcc-13.3.1/lib/gcc/x86_64-redhat-linux/13" CACHE STRING "")
 
 set(CMAKE_BUILD_TYPE "Release" CACHE STRING "")
 
@@ -21,11 +21,11 @@ set(CMAKE_BUILD_TYPE "Release" CACHE STRING "")
 #------------------------------------------------------------------------------
 if(DEFINED ENV{SPACK_CC})
 
-  set(CMAKE_C_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/none-none/compiler-wrapper-1.0-32dgmtd4fhqezj4ib4pbwjnwrxcuzxs7/libexec/spack/gcc/gcc" CACHE PATH "")
+  set(CMAKE_C_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/none-none/compiler-wrapper-1.0-32dgmtd4fhqezj4ib4pbwjnwrxcuzxs7/libexec/spack/gcc/gcc" CACHE PATH "")
 
-  set(CMAKE_CXX_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/none-none/compiler-wrapper-1.0-32dgmtd4fhqezj4ib4pbwjnwrxcuzxs7/libexec/spack/gcc/g++" CACHE PATH "")
+  set(CMAKE_CXX_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/none-none/compiler-wrapper-1.0-32dgmtd4fhqezj4ib4pbwjnwrxcuzxs7/libexec/spack/gcc/g++" CACHE PATH "")
 
-  set(CMAKE_Fortran_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/none-none/compiler-wrapper-1.0-32dgmtd4fhqezj4ib4pbwjnwrxcuzxs7/libexec/spack/gcc/gfortran" CACHE PATH "")
+  set(CMAKE_Fortran_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/none-none/compiler-wrapper-1.0-32dgmtd4fhqezj4ib4pbwjnwrxcuzxs7/libexec/spack/gcc/gfortran" CACHE PATH "")
 
 else()
 
@@ -51,11 +51,11 @@ set(MPI_CXX_COMPILER "/usr/tce/packages/mvapich2/mvapich2-2.3.7-gcc-13.3.1/bin/m
 
 set(MPI_Fortran_COMPILER "/usr/tce/packages/mvapich2/mvapich2-2.3.7-gcc-13.3.1/bin/mpif90" CACHE PATH "")
 
+set(MPIEXEC_EXECUTABLE "/usr/bin/srun" CACHE PATH "")
+
 set(MPIEXEC_NUMPROC_FLAG "-n" CACHE STRING "")
 
 set(ENABLE_MPI ON CACHE BOOL "")
-
-set(MPIEXEC_EXECUTABLE "/usr/bin/srun" CACHE PATH "")
 
 #------------------------------------------------------------------------------
 # Hardware
@@ -99,29 +99,29 @@ set(ENABLE_GTEST_DEATH_TESTS ON CACHE BOOL "")
 # TPLs
 #------------------------------------------------------------------------------
 
-set(TPL_ROOT "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_09_11_33_46/gcc-13.3.1" CACHE PATH "")
+set(TPL_ROOT "/usr/WS1/axom/libs/toss_4_x86_64_ib/2026_01_29_13_35_03/gcc-13.3.1" CACHE PATH "")
 
-set(CONDUIT_DIR "${TPL_ROOT}/conduit-0.9.5-eegw7x7gkohlehrruiegywco4xiamktk" CACHE PATH "")
+set(CONDUIT_DIR "${TPL_ROOT}/conduit-0.9.5-icskqkcmikkfw2itxiq6xdt6nilzlsvr" CACHE PATH "")
 
-set(C2C_DIR "${TPL_ROOT}/c2c-1.8.0-xcseimmxzzygacwefcb4orsdjkexiave" CACHE PATH "")
+set(C2C_DIR "${TPL_ROOT}/c2c-1.8.0-3bch3wbxcl3eji2wsbbbzmagowxe4lpc" CACHE PATH "")
 
-set(MFEM_DIR "${TPL_ROOT}/mfem-4.9.0-oeogieerz3o2odisl7obn44yzugviidm" CACHE PATH "")
+set(MFEM_DIR "${TPL_ROOT}/mfem-4.9.0-lkbiphsoccgpbaqo2cujuxcf3dqb4gga" CACHE PATH "")
 
-set(HDF5_DIR "${TPL_ROOT}/hdf5-1.8.23-dznlaqnsupnxvdgp4gvkneu32tnd2ebk" CACHE PATH "")
+set(HDF5_DIR "${TPL_ROOT}/hdf5-1.8.23-gokrordtweyfprtzgrzsycyilhy32uc7" CACHE PATH "")
 
 set(LUA_DIR "/usr" CACHE PATH "")
 
-set(RAJA_DIR "${TPL_ROOT}/raja-git.e827035c630e71a9358e2f21c2f3cf6fd5fb6605_develop-4shuz3tf6ej2nusu2p7gy3tub4fl7cit" CACHE PATH "")
+set(RAJA_DIR "${TPL_ROOT}/raja-git.3b8b59a1e9be2e1066c0d77372b3bf5956e6d6e2_develop-2u2olu3wtrgskohwge6z2r5p4pbn24xi" CACHE PATH "")
 
-set(UMPIRE_DIR "${TPL_ROOT}/umpire-git.0372fbd6e1f17d7e6dd72693f8b857f3ec7559e9_develop-mekv5ckdc5utbqgxefdhsaeuftidlnrz" CACHE PATH "")
+set(UMPIRE_DIR "${TPL_ROOT}/umpire-git.0372fbd6e1f17d7e6dd72693f8b857f3ec7559e9_develop-qqnegxhkanugllghjqvoqjilj2h4ouuf" CACHE PATH "")
 
 # OPENCASCADE not built
 
-set(ADIAK_DIR "${TPL_ROOT}/adiak-0.4.0-527775ljxopt67ho3pvm5tezd37rmjc2" CACHE PATH "")
+set(ADIAK_DIR "${TPL_ROOT}/adiak-0.4.0-45h4d32dvvatilujstudj3ucvmdeyi6a" CACHE PATH "")
 
-set(CALIPER_DIR "${TPL_ROOT}/caliper-git.527e1ae3a16487e425707e928e05c43d372e144f_master-yaa4lvtogjizlgwmk52mcunsucvykvkg" CACHE PATH "")
+set(CALIPER_DIR "${TPL_ROOT}/caliper-git.9ab0e64d2e328d1fdff2ffebb53f30a899d6a598_master-ss6he6vnxctdifczfduoqdpfyjltt2r3" CACHE PATH "")
 
-set(CAMP_DIR "${TPL_ROOT}/camp-git.a8caefa9f4c811b1a114b4ed2c9b681d40f12325_main-idirlfyyjti4dfn4lbutnn6tvabsyyzm" CACHE PATH "")
+set(CAMP_DIR "${TPL_ROOT}/camp-git.a8caefa9f4c811b1a114b4ed2c9b681d40f12325_main-idk67733aarmoow76xcm5idyfijqjntn" CACHE PATH "")
 
 # scr not built