+++ src/axom/primal/operators/detail/clip_impl.hpp
@@ -234,7 +234,7 @@ AXOM_HOST_DEVICE void poly_clip_vertices(Polyhedron<T, NDIMS>& poly,
           // A probable AMD GPU 6.3 compiler bug caused the first assert to fail.
           // The second assert is equivalent because we expect addVertex call to
           // consume the next index and increment the number of vertices.
-          // However, when the compilers no longar cause false failures, we should
+          // However, when the compilers no longer cause false failures, we should
           // restore the first assert.
           // SLIC_ASSERT(newVertexIndex == expectedVertexIndex);
           SLIC_ASSERT(newVertexIndex == poly.numVertices() - 1);