+++ src/examples/shaping_tutorial/lesson_03/klee_operators_and_validation.cpp
@@ -15,6 +15,7 @@
 #include "axom/core.hpp"
 #include "axom/primal.hpp"
 #include "axom/klee.hpp"
+#include "axom/slic.hpp"
 
 #include "axom/fmt.hpp"
 #include "axom/CLI11.hpp"